requirements.txt
python-dotenv
langchain_openai
langchain_core